Update scalafmt-core to 2.7.5 (#57)
* Update scalafmt-core to 2.7.5

* Reformat with scalafmt 2.7.5

* fix extra bracket

* fix unit tests

* more formatting

* apply scalafmtsbt

+++ janitor/src/main/scala/com/broadinstitute/dsp/janitor/NodepoolRemover.scala
@@ -19,15 +19,16 @@ object NodepoolRemover {
       override def dependencies: CheckRunnerDeps[F] = deps.checkRunnerDeps
       override def resourceToScan: fs2.Stream[F, Nodepool] = dbReader.getNodepoolsToDelete
 
-      override def checkResource(n: Nodepool, isDryRun: Boolean)(
-        implicit ev: Ask[F, TraceId]
+      override def checkResource(n: Nodepool, isDryRun: Boolean)(implicit
+        ev: Ask[F, TraceId]
       ): F[Option[Nodepool]] =
         for {
           ctx <- ev.ask
-          _ <- if (!isDryRun) {
-            val msg = DeleteNodepoolMeesage(n.nodepoolId, n.googleProject, Some(ctx))
-            deps.publisher.publishOne(msg)
-          } else F.unit
+          _ <-
+            if (!isDryRun) {
+              val msg = DeleteNodepoolMeesage(n.nodepoolId, n.googleProject, Some(ctx))
+              deps.publisher.publishOne(msg)
+            } else F.unit
         } yield Some(n)
     }
 }
